+++ CGALib/src/ThirdPersonCamera.cpp
@@ -19,10 +19,10 @@ void ThirdPersonCamera::mouseMoveCamera(float xoffset, float yoffset, float dt){
     pitch += yoffset * cameraSpeed;
     // Calculate Angle Arround
     angleAroundTarget -= xoffset * cameraSpeed;
-    if(pitch > M_PI / 2)
-        pitch = M_PI / 2 - 0.01;
-    if(pitch < -M_PI / 2)
-        pitch = -M_PI / 2 + 0.01;
+    if(pitch > static_cast<float>(M_PI) / 2)
+        pitch = static_cast<float>(M_PI) / 2 - 0.01f;
+    if(pitch < -static_cast<float>(M_PI) / 2)
+        pitch = -static_cast<float>(M_PI) / 2 + 0.01f;
     updateCamera();
 }
